hooks added for jwt authorization

+++ src/api/auth/index.ts
@@ -2,6 +2,7 @@ import dotenv from 'dotenv'
 import jwt from 'jsonwebtoken'
 
 import { FastifyPluginAsync } from 'fastify'
+import { verifyAccessToken } from '../../lib/jwtToken'
 import { LoginInput } from '../../models/DataTypes'
 
 import UserRepository from '../../repositories/UserRepository'
@@ -12,6 +13,7 @@ dotenv.config()
 const jwtKey = process.env.JWT_KEY as string
 
 const authRoute: FastifyPluginAsync = async (fastify) => {
+  let me = undefined
   fastify.post<{ Body: LoginInput }>('/signup', async (request, reply) => {
     const result = await userRepo.register(request.body)
     if (!result) {
@@ -31,11 +33,10 @@ const authRoute: FastifyPluginAsync = async (fastify) => {
   })
 
   fastify.delete('/unregister', (request, reply) => {
-    try {
-      const me = jwt.verify(request.cookies.token, jwtKey)
+    if (!(me === undefined)) {
       return userRepo.delete(me)
-    } catch (err) {
-      reply.code(401) // unauthorized
+    } else {
+      reply.code(401)
       throw new Error('unauthorized')
     }
   })
@@ -43,6 +44,12 @@ const authRoute: FastifyPluginAsync = async (fastify) => {
   fastify.post('/logout', (request, reply) => {
     reply.clearCookie('token')
     return 'logout'
+  })
+
+  fastify.addHook<{ Body: LoginInput }>('onRequest', async (request, reply) => {
+    if (!(request.cookies.token === undefined)) {
+      me = verifyAccessToken(request.cookies.token)
+    }
   })
 }
